tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/module/instantiation-error-4.html

<!DOCTYPE html>
<title>Handling of instantiation errors, 4</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    window.log = [];
    const test_load = async_test(
        "Test that loading a graph in which a module is already " +
        "errored results in an error.");

    window.addEventListener("error", ev => {
      test_load.step(() => assert_equals(ev.error.constructor, SyntaxError));
      log.push(ev.message);
    });

    window.addEventListener("load", test_load.step_func_done(ev => {
      const msg = log[0];
      assert_array_equals(log, [msg, 1, msg, 2]);
    }));

    function unreachable() { log.push("unexpected"); }
</script>
<script type="module" src="./instantiation-error-4a.js"
    onerror="unreachable()" onload="log.push(1)"></script>
<script type="module" src="./instantiation-error-4d.js"
    onerror="unreachable()" onload="log.push(2)"></script>
